refactor : refactor name of some files
