+++ src/application/use_cases/comment/create-comment-use-case.ts
@@ -1,8 +1,8 @@
 import { PrismaCommentRepository } from '../../../infra/database/prisma/prisma_repositories/prisma-comment-repository';
-import { GetUserByIdUseCase } from '../user/getUserUseCase';
+import { GetUserByIdUseCase } from '../user/get-user-use-case';
 import { CommentRequest, CommentResponse } from '../../../utils/@interfaces';
 import { MissingParamError, ParamDoesNotExist } from '../../../utils/errors/index.';
-import { prisma } from '../../../infra/database/prisma/client/prismaClient';
+import { prisma } from '../../../infra/database/prisma/client/prisma-client';
 import { logger } from '../../../utils/logger';
 
 export class CreateCommentUseCase {
