+++ src/application/use_cases/factories/comment/make-create-comment-use-case.ts
@@ -1,7 +1,7 @@
 import { PrismaCommentRepository } from '../../../../infra/database/prisma/prisma_repositories/prisma-comment-repository';
 import { PrismaUserRepository } from '../../../../infra/database/prisma/prisma_repositories/prisma-user-repository';
 import { CreateCommentUseCase } from '../../comment/create-comment-use-case';
-import { GetUserByIdUseCase } from '../../user/getUserUseCase';
+import { GetUserByIdUseCase } from '../../user/get-user-use-case';
 
 export async function makeCreateComment() {
   const createCommentRepository = new PrismaCommentRepository();
