+++ src/application/use_cases/factories/user/make-get-user-use-case.ts
@@ -1,4 +1,4 @@
-import { GetUserByIdUseCase } from '../../user/getUserUseCase';
+import { GetUserByIdUseCase } from '../../user/get-user-use-case';
 import { PrismaUserRepository } from '../../../../infra/database/prisma/prisma_repositories/prisma-user-repository';
 
 export async function makeGetUserUseCase() {
