+++ src/infra/database/prisma/prisma_repositories/prisma-user-repository.ts
@@ -1,6 +1,6 @@
 import { User, Prisma } from '@prisma/client';
 import { UserRepository } from '../../../../application/repositories/user-repository';
-import { prisma } from '../client/prismaClient';
+import { prisma } from '../client/prisma-client';
 import { logger } from '../../../../utils/logger';
 
 export class PrismaUserRepository implements UserRepository {
